requirements.txt
django==2.0.10
djangorestframework==3.9.0
django-rest-swagger==2.2.0
drf-yasg==1.10.1
coverage==4.5.1
pytest==3.8.0
model-mommy==1.6.0
short_url==1.2.2